🚚 Moved meta components up to `/components` level
