+++ src/pages/tutorials.js
@@ -5,7 +5,7 @@ import { graphql } from 'gatsby'
 import Layout from '../components/layouts/default'
 import PostCard from '../components/postcard'
 import { Spirit } from '../components/spirit-styles'
-import { MetaData } from '../components/layouts/partials/meta'
+import { MetaData } from '../components/meta'
 // import SectionHeading from '../components/layouts/partials/section-heading'
 
 class TutorialsPage extends React.Component {
