+++ src/templates/tags-faq.js
@@ -6,7 +6,7 @@ import Link from 'gatsby-link'
 import Layout from '../components/layouts/default'
 import FAQ from '../components/faq'
 import { Spirit } from '../components/spirit-styles'
-import { MetaData } from '../components/layouts/partials/meta'
+import { MetaData } from '../components/meta'
 import FAQTagList from '../components/layouts/partials/faq-taglist'
 
 class FAQTags extends React.Component {
